requirements.txt
Flask 
flask_jwt_extended 
flask_bcrypt 
python-dotenv 
mysql-connector-python 
Flask-Cors 
